requirements.txt
accelerate==0.17.1
asttokens==2.2.1
backcall==0.2.0
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.0.7
controlnet-aux==0.0.1
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
diffusers==0.15.0.dev0
einops==0.6.0
executing==1.2.0
filelock==3.9.0
flatbuffers==23.3.3
flit_core==3.8.0
fonttools==4.39.2
huggingface-hub==0.13.3
humanfriendly==10.0
idna==3.4
imageio==2.26.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==2.8.4
numpy==1.23.5
onnxruntime==1.14.1
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyre-extensions==0.0.23
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.1
scikit-image==0.20.0
scipy==1.9.1
setuptools==65.6.3
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tifffile==2023.3.21
tokenizers==0.13.2
torch==2.0.0+cu117
torchaudio==2.0.0
torchvision==0.15.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.3
typing_extensions==4.4.0
typing-inspect==0.8.0
urllib3==1.26.14
wcwidth==0.2.6
wheel==0.38.4
win-inet-pton==1.1.0
wincertstore==0.2
xformers==0.0.16
zipp==3.15.0
